Background Service template added

+++ Android/app/src/main/java/com/amit/esp32_trackingwithservoapp/Interfaces/IMyRotationVector.java
@@ -1,5 +1,5 @@
 package com.amit.esp32_trackingwithservoapp.Interfaces;
 
 public interface IMyRotationVector {
-    public void onNewRotationVectorValuesAvaible(float x);
+    public void onNewRotationVectorValuesAvaible(double x);
 }
